+++ pkg/netconf/response.go
@@ -49,16 +49,27 @@ func NewDataReply(messageID string, data []byte) *RPCReply {
 func NewErrorReply(messageID string, err *RPCError) *RPCReply {
 	return &RPCReply{
 		MessageID: messageID,
-		Errors:    []*RPCError{err},
+		Errors:    []*RPCError{normalizeRPCError(err)},
 	}
 }
 
 // NewMultiErrorReply creates an error <rpc-reply> with multiple <rpc-error>
 func NewMultiErrorReply(messageID string, errors []*RPCError) *RPCReply {
+	normalized := make([]*RPCError, len(errors))
+	for i, err := range errors {
+		normalized[i] = normalizeRPCError(err)
+	}
 	return &RPCReply{
 		MessageID: messageID,
-		Errors:    errors,
+		Errors:    normalized,
 	}
+}
+
+func normalizeRPCError(err *RPCError) *RPCError {
+	if err != nil {
+		return err
+	}
+	return ErrOperationFailed("rpc error unavailable")
 }
 
 func (r *RPCReply) WithAttributes(attrs []xml.Attr) *RPCReply {